+++ examples/data-sources/meraki_appliance_prefix_delegated_static/data-source.tf
@@ -1,0 +1,4 @@
+data "meraki_appliance_prefix_delegated_static" "example" {
+  id         = "12345678"
+  network_id = "L_123456"
+}