+++ examples/data-sources/meraki_appliance_prefix_delegated_statics/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_prefix_delegated_statics" "example" {
+  network_id = "L_123456"
+}