+++ examples/resources/meraki_appliance_prefix_delegated_static/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_prefix_delegated_static.example "<network_id>,<id>"