+++ docs/guides/changelog.md
@@ -20,6 +20,8 @@ description: |-
 - Add `meraki_appliance_firewall_settings` resource and data source
 - Add `meraki_appliance_port` resource and data source
 - Add `meraki_appliance_vlans_settings` resource and data source
+- Add `meraki_appliance_prefix_delegated_static` resource and data source
+- Add `meraki_appliance_prefix_delegated_statics` data source
 
 ## 0.1.1